smartprocesspool_examples/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartprocesspool-examples"
version = "0.1.0"
description = "Examples for smartprocesspool."
readme = "README.md"
authors = [
    { name = "王炳辉", email = "binghui.wang@foxmail.com" }
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "smartprocesspool",
    "matplotlib",
    "scikit-learn",
    "numpy",
    "rich",
    "torchvision",
    "click"
]

[project.urls]
Homepage = "https://github.com/Time-Coder/smartprocesspool"
Repository = "https://github.com/Time-Coder/smartprocesspool.git"